+++ graphics/src/ColladaLoader.cc
@@ -926,8 +926,14 @@ void ColladaLoaderPrivate::LoadAnimationSet(tinyxml2::XMLElement *_xml,
         frameTransXml->FirstChildElement("technique_common");
       accessor = accessor->FirstChildElement("accessor");
 
-      unsigned int stride =
-        ignition::math::parseInt(accessor->Attribute("stride"));
+      // stride is optional, default to 1
+      unsigned int stride = 1;
+      auto *strideAttribute = accessor->Attribute("stride");
+      if (strideAttribute)
+      {
+        stride = static_cast<unsigned int>(
+            ignition::math::parseInt(strideAttribute));
+      }
 
       SkeletonNode *targetNode = _skel->NodeById(targetBone);
       if (targetNode == nullptr)